changed signiture of http.AttachProfiler, moved GorillaMux adapter to separete package
