+++ http/http.go
@@ -31,7 +31,7 @@ type stats struct {
 }
 
 func init() {
-	AttachProfiler(http.DefaultServeMux, true)
+	http.HandleFunc("/debug/pprofstats", Stats)
 }
 
 // router
@@ -40,23 +40,21 @@ type router interface {
 	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
 }
 
-// AttachProfiler will register http profiling routes to http router (http.ServeMux like type that satisfy router interface)
-// If you are not using http.DefaultServerMux set pprofRegistrated to false so it's http routes can also be registered.
-func AttachProfiler(r router, pprofRegistrated bool) {
+// AttachProfiler will register http profiling routes to http router (http.ServeMux like type that satisfy router interface).
+// Calling AttachProfiler is needed if you are not using http.DefaultServeMux.
+func AttachProfiler(r router) {
 	r.HandleFunc("/debug/pprofstats", Stats)
 
-	if !pprofRegistrated {
-		r.HandleFunc("/debug/pprof/", httppprof.Index)
-		r.HandleFunc("/debug/pprof/cmdline", httppprof.Cmdline)
-		r.HandleFunc("/debug/pprof/profile", httppprof.Profile)
-		r.HandleFunc("/debug/pprof/symbol", httppprof.Symbol)
-		r.HandleFunc("/debug/pprof/trace", httppprof.Trace)
+	r.HandleFunc("/debug/pprof/", httppprof.Index)
+	r.HandleFunc("/debug/pprof/cmdline", httppprof.Cmdline)
+	r.HandleFunc("/debug/pprof/profile", httppprof.Profile)
+	r.HandleFunc("/debug/pprof/symbol", httppprof.Symbol)
+	r.HandleFunc("/debug/pprof/trace", httppprof.Trace)
 
-		r.Handle("/debug/pprof/goroutine", httppprof.Handler("goroutine"))
-		r.Handle("/debug/pprof/heap", httppprof.Handler("heap"))
-		r.Handle("/debug/pprof/threadcreate", httppprof.Handler("threadcreate"))
-		r.Handle("/debug/pprof/block", httppprof.Handler("block"))
-	}
+	r.Handle("/debug/pprof/goroutine", httppprof.Handler("goroutine"))
+	r.Handle("/debug/pprof/heap", httppprof.Handler("heap"))
+	r.Handle("/debug/pprof/threadcreate", httppprof.Handler("threadcreate"))
+	r.Handle("/debug/pprof/block", httppprof.Handler("block"))
 }
 
 // Stats exposes pprof status counters, includes number of goroutines, threads, blocks
